Move crc16 function to libs
