+++ Marlin/src/core/utility.cpp
@@ -35,19 +35,6 @@ void safe_delay(millis_t ms) {
   thermalManager.manage_heater(); // This keeps us safe if too many small safe_delay() calls are made
 }
 
-#if EITHER(EEPROM_SETTINGS, SD_FIRMWARE_UPDATE)
-
-  void crc16(uint16_t *crc, const void * const data, uint16_t cnt) {
-    uint8_t *ptr = (uint8_t *)data;
-    while (cnt--) {
-      *crc = (uint16_t)(*crc ^ (uint16_t)(((uint16_t)*ptr++) << 8));
-      for (uint8_t i = 0; i < 8; i++)
-        *crc = (uint16_t)((*crc & 0x8000) ? ((uint16_t)(*crc << 1) ^ 0x1021) : (*crc << 1));
-    }
-  }
-
-#endif // EEPROM_SETTINGS || SD_FIRMWARE_UPDATE
-
 #if ENABLED(DEBUG_LEVELING_FEATURE)
 
   #include "../module/probe.h"
