+++ Marlin/src/feature/backlash.h
@@ -24,6 +24,8 @@
 #include "../inc/MarlinConfigPre.h"
 #include "../module/planner.h"
 
+constexpr uint8_t all_on = 0xFF, all_off = 0x00;
+
 class Backlash {
 public:
   #ifdef BACKLASH_DISTANCE_MM
